+++ src/pytileproj/tile.py
@@ -656,7 +656,6 @@ class RasterTile(BaseModel, Generic[T_co]):
         x: float,
         y: float,
         crs: Any | None = None,  # noqa: ANN401
-        px_origin: OriginStr | None = None,
     ) -> tuple[int, int]: ...
 
     @overload
@@ -665,7 +664,6 @@ class RasterTile(BaseModel, Generic[T_co]):
         x: npt.NDArray[Any],
         y: float,
         crs: Any | None = None,  # noqa: ANN401
-        px_origin: OriginStr | None = None,
     ) -> tuple[npt.NDArray[Any], int]: ...
 
     @overload
@@ -674,7 +672,6 @@ class RasterTile(BaseModel, Generic[T_co]):
         x: float,
         y: npt.NDArray[Any],
         crs: Any | None = None,  # noqa: ANN401
-        px_origin: OriginStr | None = None,
     ) -> tuple[int, npt.NDArray[Any]]: ...
 
     @overload
@@ -683,7 +680,6 @@ class RasterTile(BaseModel, Generic[T_co]):
         x: npt.NDArray[Any],
         y: npt.NDArray[Any],
         crs: Any | None = None,  # noqa: ANN401
-        px_origin: OriginStr | None = None,
     ) -> tuple[npt.NDArray[Any], npt.NDArray[Any]]: ...
 
     def xy2rc(
@@ -691,7 +687,6 @@ class RasterTile(BaseModel, Generic[T_co]):
         x: float | npt.NDArray[Any],
         y: float | npt.NDArray[Any],
         crs: Any | None = None,
-        px_origin: OriginStr | None = None,
     ) -> tuple[int | npt.NDArray[Any], int | npt.NDArray[Any]]:
         """Convert world system to pixels coordinates.
 
@@ -728,8 +723,7 @@ class RasterTile(BaseModel, Generic[T_co]):
         """
         if crs is not None:
             x, y = transform_coords(x, y, crs, self.crs)
-        px_origin = self.px_origin if px_origin is None else px_origin
-        c, r = xy2ij(x, y, self.geotrans, origin=px_origin)
+        c, r = xy2ij(x, y, self.geotrans)
         return r, c
 
     @overload
@@ -1001,9 +995,21 @@ class RasterTile(BaseModel, Generic[T_co]):
         """
         return not self == other
 
+    def __repr__(self) -> str:
+        """Short string representation."""
+        return f"{self.__class__.__name__}({self.name})"
+
     def __str__(self) -> str:
-        """Representation of a raster tile as a Well Known Text (WKT) string."""
-        return self.boundary_wkt
+        """Extensive string representation."""
+        n_chars = len(self.__class__.__name__)
+        return (
+            f"{self.__class__.__name__} \n{'-' * n_chars} \n"
+            f"Name: \n{self.name} \n"
+            f"Shape: \n({self.n_rows}, {self.n_cols})\n"
+            f"Projection: \n{self.pyproj_crs.to_proj4()}\n"
+            f"Geotransformation parameters: \n{self.geotrans}\n"
+            f"Pixel origin: \n{self.px_origin}"
+        )
 
 
 if __name__ == "__main__":